pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secure-proxy-gateway"
version = "1.0.0"
description = "Lightweight configurable secure proxy gateway"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.109.0",
  "uvicorn[standard]>=0.27.0",
  "httpx>=0.26.0",
  "pydantic>=2.5.0",
  "pyyaml>=6.0.1",
  "typer[all]>=0.9.0",
  "jinja2>=3.1.2",
  "rich>=13.7.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
